+++ packages/page-assets/src/index.tsx
@@ -5,28 +5,35 @@
 import '@polkadot/api-augment/substrate';
 
 import type { BN } from '@polkadot/util';
+import type { HexString } from '@polkadot/util/types';
 
 import React, { useMemo, useRef } from 'react';
 import { Route, Routes } from 'react-router';
 
+import { getGenesis } from '@polkadot/apps-config';
 import { Tabs } from '@polkadot/react-components';
-import { useAccounts } from '@polkadot/react-hooks';
+import { useAccounts, useApi, useAssetIds, useAssetInfos } from '@polkadot/react-hooks';
 import { BN_ONE } from '@polkadot/util';
 
 import Balances from './Balances/index.js';
 import Overview from './Overview/index.js';
 import { useTranslation } from './translate.js';
-import useAssetIds from './useAssetIds.js';
-import useAssetInfos from './useAssetInfos.js';
 
 interface Props {
   basePath: string;
   className?: string;
 }
 
-function findOpenId (ids?: BN[]): BN {
+// Chains in which next asset id should be incremented from 1
+const GENESIS_HASHES = [getGenesis('statemint'), getGenesis('statemine')];
+
+function findOpenId (genesisHash: HexString, ids?: BN[]): BN {
   if (!ids?.length) {
     return BN_ONE;
+  }
+
+  if (GENESIS_HASHES.includes(genesisHash)) {
+    return ids.sort((a, b) => a.cmp(b))[ids.length - 1].add(BN_ONE);
   }
 
   const lastTaken = ids.find((id, index) =>
@@ -36,12 +43,13 @@ function findOpenId (ids?: BN[]): BN {
   );
 
   return lastTaken
-    ? lastTaken.sub(BN_ONE)
+    ? lastTaken.add(BN_ONE)
     : ids[ids.length - 1].add(BN_ONE);
 }
 
 function AssetApp ({ basePath, className }: Props): React.ReactElement<Props> {
   const { t } = useTranslation();
+  const { api } = useApi();
   const { hasAccounts } = useAccounts();
   const ids = useAssetIds();
   const infos = useAssetInfos(ids);
@@ -66,8 +74,8 @@ function AssetApp ({ basePath, className }: Props): React.ReactElement<Props> {
   );
 
   const openId = useMemo(
-    () => findOpenId(ids),
-    [ids]
+    () => findOpenId(api.genesisHash.toHex(), ids),
+    [api.genesisHash, ids]
   );
 
   return (